+++ common/Storyboarding/Util/OsbAnimationPool.cs
@@ -5,10 +5,10 @@ namespace StorybrewCommon.Storyboarding.Util
     public class OsbAnimationPool : OsbSpritePool
     {
         private int frameCount;
-        private int frameDelay;
+        private double frameDelay;
         private OsbLoopType loopType;
 
-        public OsbAnimationPool(StoryboardLayer layer, string path, int frameCount, int frameDelay, OsbLoopType loopType, OsbOrigin origin, Action<OsbSprite, double, double> finalizeSprite = null)
+        public OsbAnimationPool(StoryboardLayer layer, string path, int frameCount, double frameDelay, OsbLoopType loopType, OsbOrigin origin, Action<OsbSprite, double, double> finalizeSprite = null)
             : base(layer, path, origin, finalizeSprite)
         {
             this.frameCount = frameCount;
@@ -16,7 +16,7 @@ namespace StorybrewCommon.Storyboarding.Util
             this.loopType = loopType;
         }
 
-        public OsbAnimationPool(StoryboardLayer layer, string path, int frameCount, int frameDelay, OsbLoopType loopType, OsbOrigin origin, bool additive)
+        public OsbAnimationPool(StoryboardLayer layer, string path, int frameCount, double frameDelay, OsbLoopType loopType, OsbOrigin origin, bool additive)
             : this(layer, path, frameCount, frameDelay, loopType, origin, additive ? (sprite, startTime, endTime) => sprite.Additive(startTime, endTime) : (Action<OsbSprite, double, double>)null)
         {
         }
